refactor components to files
