+++ src/js/icons/Arrow.tsx
@@ -1,0 +1,11 @@
+import React from "react"
+
+const Arrow = (props) => {
+  return (
+    <svg className={props.className} onClick={props.onClick} viewBox="0 0 8 8">
+      <polygon points="0 0 8 4 0 8" />
+    </svg>
+  )
+}
+
+export default Arrow
